requirements.txt
pickledb==0.7.2
simplejson==3.16.0
pyyaml==3.13
boto3==1.9.12
botocore==1.12.12
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.3
s3transfer==0.1.13
urllib3==1.23
pytz==2018.5
colorlog==4.0.2
xxhash==1.3.0
